Refactor VerseList component into hooks and sub-components
- Extract audio logic to `useQuranAudio` hook
- Extract settings logic to `useQuranSettings` hook
- Extract tafsir logic to `useTafsir` hook
- Create `VerseItem`, `MushafView`, `AudioPlayer`, `SettingsDialog` components
- Create `src/types/quran.ts` for shared types
- Move utility functions to `src/components/quran/utils.ts`
- Update `SurahList`, `QuranBrowser`, `kemenag-api` to use shared types
- Improve code organization and maintainability of `VerseList.tsx`
